pyproject.toml
[project]
name = "adk-blueprints"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-community[gmail]>=2.0.7",
    "langchain-google-genai>=2.1.8",
]
